# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  client: 'client',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  role: UserProfileRole;
  createdAt: string;
}

export interface UserProfileUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
}

export type RoleUpdateRole = typeof RoleUpdateRole[keyof typeof RoleUpdateRole];


export const RoleUpdateRole = {
  client: 'client',
  admin: 'admin',
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
}

export type LegalQueryStatus = typeof LegalQueryStatus[keyof typeof LegalQueryStatus];


export const LegalQueryStatus = {
  pending: 'pending',
  in_review: 'in_review',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface LegalQuery {
  id: number;
  userId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  subject: string;
  description: string;
  category: string;
  status: LegalQueryStatus;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type LegalQueryInputCategory = typeof LegalQueryInputCategory[keyof typeof LegalQueryInputCategory];


export const LegalQueryInputCategory = {
  family_law: 'family_law',
  criminal_law: 'criminal_law',
  corporate_law: 'corporate_law',
  property_law: 'property_law',
  employment_law: 'employment_law',
  immigration_law: 'immigration_law',
  other: 'other',
} as const;

export interface LegalQueryInput {
  /** @minLength 3 */
  subject: string;
  /** @minLength 10 */
  description: string;
  category: LegalQueryInputCategory;
}

export type LegalQueryUpdateStatus = typeof LegalQueryUpdateStatus[keyof typeof LegalQueryUpdateStatus];


export const LegalQueryUpdateStatus = {
  pending: 'pending',
  in_review: 'in_review',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface LegalQueryUpdate {
  status?: LegalQueryUpdateStatus;
  adminNotes?: string;
}

export interface Document {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  fileUrl: string;
  fileName?: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}

export interface DocumentInput {
  title: string;
  description?: string;
  category: string;
  fileUrl: string;
  fileName: string;
  fileSize?: number;
}

export interface DashboardStats {
  totalQueries: number;
  pendingQueries: number;
  inReviewQueries: number;
  resolvedQueries: number;
  totalDocuments: number;
  totalClients: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ContactInfo {
  firmName: string;
  address: string;
  city?: string;
  state?: string;
  zip?: string;
  phone: string;
  email: string;
  hours: string;
  /** @nullable */
  mapEmbedUrl?: string | null;
}

export type ListUsersParams = {
role?: ListUsersRole;
};

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  client: 'client',
  admin: 'admin',
} as const;

export type ListQueriesParams = {
status?: ListQueriesStatus;
category?: string;
};

export type ListQueriesStatus = typeof ListQueriesStatus[keyof typeof ListQueriesStatus];


export const ListQueriesStatus = {
  pending: 'pending',
  in_review: 'in_review',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export type ListDocumentsParams = {
category?: string;
};