# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legal Services Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: queries
    description: Client legal queries
  - name: documents
    description: Admin document uploads
  - name: users
    description: User profile and role management
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [client, admin]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /queries:
    get:
      operationId: listQueries
      tags: [queries]
      summary: List queries (admin sees all, client sees own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_review, resolved, closed]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of queries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalQuery"
    post:
      operationId: createQuery
      tags: [queries]
      summary: Submit a new legal query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalQueryInput"
      responses:
        "201":
          description: Created query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalQuery"

  /queries/{queryId}:
    get:
      operationId: getQuery
      tags: [queries]
      summary: Get a specific query
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Query details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalQuery"
    patch:
      operationId: updateQuery
      tags: [queries]
      summary: Update query status or add admin notes (admin only)
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalQueryUpdate"
      responses:
        "200":
          description: Updated query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalQuery"
    delete:
      operationId: deleteQuery
      tags: [queries]
      summary: Delete a query (admin only)
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document (admin only)
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /stats/queries-by-category:
    get:
      operationId: getQueriesByCategory
      tags: [stats]
      summary: Get query count breakdown by category
      responses:
        "200":
          description: Queries by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"

  /contact:
    get:
      operationId: getContactInfo
      tags: [users]
      summary: Get firm contact information and location
      responses:
        "200":
          description: Contact info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactInfo"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UserProfile:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, admin]
        createdAt:
          type: string

    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string

    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [client, admin]

    LegalQuery:
      type: object
      required: [id, userId, subject, description, category, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        subject:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, in_review, resolved, closed]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    LegalQueryInput:
      type: object
      required: [subject, description, category]
      properties:
        subject:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        category:
          type: string
          enum: [family_law, criminal_law, corporate_law, property_law, employment_law, immigration_law, other]

    LegalQueryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_review, resolved, closed]
        adminNotes:
          type: string

    Document:
      type: object
      required: [id, title, category, fileUrl, uploadedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: ["integer", "null"]
        uploadedAt:
          type: string

    DocumentInput:
      type: object
      required: [title, category, fileUrl, fileName]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer

    DashboardStats:
      type: object
      required: [totalQueries, pendingQueries, inReviewQueries, resolvedQueries, totalDocuments, totalClients]
      properties:
        totalQueries:
          type: integer
        pendingQueries:
          type: integer
        inReviewQueries:
          type: integer
        resolvedQueries:
          type: integer
        totalDocuments:
          type: integer
        totalClients:
          type: integer

    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    ContactInfo:
      type: object
      required: [firmName, address, phone, email, hours]
      properties:
        firmName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: string
        email:
          type: string
        hours:
          type: string
        mapEmbedUrl:
          type: ["string", "null"]