# lib/api-zod/src/generated/types/contactInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInfo {
  firmName: string;
  address: string;
  city?: string;
  state?: string;
  zip?: string;
  phone: string;
  email: string;
  hours: string;
  /** @nullable */
  mapEmbedUrl?: string | null;
}