# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  fileUrl: string;
  fileName?: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}