# lib/api-zod/src/generated/types/legalQuery.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LegalQueryStatus } from './legalQueryStatus';

export interface LegalQuery {
  id: number;
  userId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  subject: string;
  description: string;
  category: string;
  status: LegalQueryStatus;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}