# lib/api-zod/src/generated/types/legalQueryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LegalQueryInputCategory } from './legalQueryInputCategory';

export interface LegalQueryInput {
  /** @minLength 3 */
  subject: string;
  /** @minLength 10 */
  description: string;
  category: LegalQueryInputCategory;
}