# lib/api-zod/src/generated/types/legalQueryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Legal Services Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LegalQueryUpdateStatus } from './legalQueryUpdateStatus';

export interface LegalQueryUpdate {
  status?: LegalQueryUpdateStatus;
  adminNotes?: string;
}